requirements.txt
py-cord>=2.7.0
python-dotenv==1.0.1
PyNaCl==1.5.0
audioop-lts==0.2.1
